+++ samples.h
@@ -30,7 +30,7 @@ typedef struct
 int16_t sample_get(int16_t index);
 void sample_set(int16_t index, int16_t value);
 int16_t sample_get_interpolated(float index);
-uint32_t sample_get_average(uint16_t nSamples);
+uint16_t sample_get_average(uint16_t nSamples);
 void sample_clear_average(void);
 
 #endif
